fix(node:stream): split JSON helpers out of readwrite to satisfy file-size gate
node_stream_readwrite.rs reached 2001 lines on main, one over the
2000-line lint gate (scripts/check_file_size.sh), turning lint red for
every open PR. Extract the self-contained push_json_number +
try_stringify_node_stream_json helpers into node_stream_json.rs
(re-globbed via pub use json_stream::*), bringing the file to 1952.
